refactor: reorganize and document CSG
