+++ primap2/csg/__init__.py
@@ -3,3 +3,15 @@
 Generate a composite harmonized dataset from multiple sources according to defined
 source priorities and matching algorithms.
 """
+
+from ._compose import compose
+from ._models import PriorityDefinition, StrategyDefinition, TimeseriesSelector
+from ._strategies.substitution import SubstitutionStrategy
+
+__all__ = [
+    "compose",
+    "PriorityDefinition",
+    "StrategyDefinition",
+    "TimeseriesSelector",
+    "SubstitutionStrategy",
+]
