+++ primap2/csg/_compose.py
@@ -1,232 +1,137 @@
 """Compose a harmonized dataset from multiple input datasets."""
 
 import math
-import typing
 from collections.abc import Hashable
 
 import numpy as np
 import tqdm
 import xarray as xr
-from attrs import define
 from loguru import logger
 
+from . import _models
 
-@define
-class PriorityDefinition:
+
+def compose(
+    *,
+    input_data: xr.Dataset,
+    priority_definition: _models.PriorityDefinition,
+    strategy_definition: _models.StrategyDefinition,
+    progress_bar: bool = True,
+) -> xr.Dataset:
     """
-    Defines source priorities for composing a full dataset or a single timeseries.
+    Compose a harmonized dataset from multiple input datasets.
 
-    Attributes
+    The input datasets are treated at the timeseries level, for each timeseries:
+
+    * the highest priority dataset is chosen according to the priority_definition
+    * if the dataset contains any nan values (denoting missing information), the next
+      highest priority dataset is selected to fill the missing values
+    * a filling strategy is chosen according to the strategy_definition and the filling
+      is done with the dataset selected in the previous step
+    * if afterwards there are still missing nan values, the next highest priority
+      dataset is selected and the previous step is repeated
+    * when all missing values are filled or all input datasets are exhausted, the
+      timeseries is finished and written to the result.
+
+    In addition to the harmonized data, also a description of the processing steps
+    done for each timeseries is returned in the result dataset, where for each
+    entity, a variable of the form "Processing of $entity" is returned, with the same
+    dimensions as the entity, apart from the time dimension.
+
+    Parameters
     ----------
-    selection_dimensions
-        List of dimensions from which source timeseries are selected. These are the
-        priority dimensions, and each priority has to specify all selection dimensions.
-    priorities
-        List of priority selections. Higher priority selections come first. Each
-        selection consists of a dict which maps dimension names to values. Each
-        selection has to specify all selection_dimensions, but may specify additional
-        dimensions (fixed dimensions) to limit the selection to specific cases.
-        Examples:
-        [{"area (ISO3)": "COL", "source": "A"}, {"source": "B"}]
-        would select source "A" for Columbia, but source "B" for all other countries.
-        In this case, selection_dimensions = ["source"].
+    input_data
+        Dataset with the input data. The input data dimensions determine the output
+        data dimensions: the output data has the same dimensions as the input data,
+        minus the priority dimensions defined in the priority_definition. From the
+        priority dimensions, the different datasets for the filling are selected, so
+        they vanish in the result.
+    priority_definition
+        Defines the priorities to select timeseries from the input data. Priorities
+        are formed by a list of selections and are used "from left to right", where the
+        first matching selection has the highest priority. Each selection has to specify
+        values for all priority dimensions (so that exactly one timeseries is selected
+        from the input data), but can also specify other dimensions. That way it is,
+        e.g., possible to define a different priority for a specific country by listing
+        it early (i.e. with high priority) before the more general rules which should
+        be applied for all other countries.
+    strategy_definition
+        Defines the filling strategies to be used when filling timeseries with other
+        timeseries. Again, the priority is defined by a list of selections and
+        corresponding strategies which are used "from left to right". Selections can use
+        any dimension and don't have to apply to only one timeseries. For example, to
+        define a default strategy which should be used for all timeseries unless
+        something else is configured, configure an empty selection as the last
+        (rightmost) entry.
+    progress_bar
+        If True (default), show progress bars using the tqdm package during the
+        operation.
+
+    Returns
+    -------
+        result. Dataset with the same entities and dimensions as input_data, but with
+        following changes: the data is composed and filled according to the rules,
+        the priority dimensions are reduced and not included in the result, and
+        additional variables of the form "Processing of $entity" are added which
+        describe the processing steps done for each timeseries.
     """
+    result_das = {}
 
-    selection_dimensions: list[Hashable]
-    priorities: list[dict[Hashable, str]]
-
-    def limit(self, dim: Hashable, value: str) -> typing.Self:
-        """Remove one additional dimension by limiting to a single value.
-
-        You can't remove selection dimensions, only additional (fixed) dimensions.
-        """
-        new_priorities = []
-        for sel in self.priorities:
-            if dim not in sel:
-                new_priorities.append(sel)
-            elif sel[dim] == value:
-                # filter out the matching value
-                new_priorities.append({k: v for k, v in sel.items() if k != dim})
-        return PriorityDefinition(
-            selection_dimensions=self.selection_dimensions, priorities=new_priorities
+    if progress_bar:
+        entity_iterator = tqdm.tqdm(input_data)
+    else:
+        entity_iterator = input_data
+    for entity in entity_iterator:
+        if progress_bar:
+            entity_iterator.set_postfix_str(entity)
+        input_da = input_data[entity].pr.dequantify()
+        # all dimensions are either time, priority selection dimensions, or need to
+        # be iterated over
+        group_by_dimensions = tuple(
+            dim
+            for dim in input_da.dims
+            if dim != "time" and dim not in priority_definition.selection_dimensions
         )
-
-
-@define
-class ProcessingStepDescription:
-    """Structured description of a processing step done on a timeseries."""
-
-    time: np.ndarray[np.datetime64] | typing.Literal["all"]
-    processing_description: str
-
-    def __str__(self) -> str:
-        return f"In time={self.time}: {self.processing_description}"
-
-
-@define
-class TimeseriesProcessingDescription:
-    """Structured description of all processing steps done on a timeseries."""
-
-    steps: list[ProcessingStepDescription]
-
-    def __str__(self) -> str:
-        return "\n".join(str(step) for step in self.steps)
-
-
-class FillingStrategyModel(typing.Protocol):
-    """
-    Fill missing data in a timeseries using another timeseries.
-    """
-
-    type: str
-
-    def fill(
-        self,
-        *,
-        ts: xr.DataArray,
-        fill_ts: xr.DataArray,
-        fill_ts_repr: str,
-    ) -> tuple[xr.DataArray, list[ProcessingStepDescription]]:
-        """Fill gaps in ts using data from the fill_ts.
-
-        Parameters
-        ----------
-        ts
-            Base timeseries. Missing data (NaNs) in this timeseries will be filled.
-            This function must not modify the data in ts.
-        fill_ts
-            Fill timeseries. Data from this timeseries will be used (possibly after
-            modification) to fill missing data in the base timeseries.
-            This function must not modify the data in fill_ts.
-        fill_ts_repr
-            String representation of fill_ts. Human-readable short representation of
-            the fill_ts (e.g. the source).
-
-        Returns
-        -------
-            filled_ts, descriptions. filled_ts contains the result, where missing
-            data in ts is (partly) filled using information from fill_ts.
-            descriptions contains human-readable, structured descriptions of how the
-            data was processed, grouped by years for which the same processing steps
-            were taken. Every year for which data was changed has to be described and
-            no year for which data was not changed is allowed to be described.
-        """
-        ...
-
-
-class SubstitutionStrategy:
-    """Fill missing data in the result dataset by copying.
-
-    The NaNs in the result dataset are substituted with data from the filling
-    dataset.
-    """
-
-    type = "substitution"
-
-    def fill(
-        self,
-        *,
-        ts: xr.DataArray,
-        fill_ts: xr.DataArray,
-        fill_ts_repr: str,
-    ) -> tuple[xr.DataArray, list[ProcessingStepDescription]]:
-        """Fill gaps in ts using data from the fill_ts.
-
-        Parameters
-        ----------
-        ts
-            Base timeseries. Missing data (NaNs) in this timeseries will be filled.
-            This function does not modify the data in ts.
-        fill_ts
-            Fill timeseries. Data from this timeseries will be used (possibly after
-            modification) to fill missing data in the base timeseries.
-            This function does not modify the data in fill_ts.
-        fill_ts_repr
-            String representation of fill_ts. Human-readable short representation of
-            the fill_ts (e.g. the source).
-
-        Returns
-        -------
-            filled_ts, descriptions. filled_ts contains the result, where missing
-            data in ts is (partly) filled using unmodified data from fill_ts.
-            descriptions contains information about which years were affected and
-            filled how.
-        """
-        filled_ts = xr.core.ops.fillna(ts, fill_ts, join="exact")
-        filled_mask = ts.isnull() & ~fill_ts.isnull()
-        time_filled = filled_mask["time"][filled_mask].to_numpy()
-        description = ProcessingStepDescription(
-            time=time_filled,
-            processing_description="substituted with corresponding values from"
-            f" {fill_ts_repr}",
+        # pre-allocate result arrays which will be filled timeseries-by-timeseries
+        result_dimensions = ["time", *group_by_dimensions]
+        result_das[entity] = xr.DataArray(
+            data=np.nan,
+            dims=result_dimensions,
+            coords=[input_da.coords[dim] for dim in result_dimensions],
         )
-        return filled_ts, [description]
-
-
-@define
-class TimeseriesSelector:
-    selections: dict[Hashable, str | list[str]]
-    """
-    Examples:
-    {"source": "FAOSTAT", "scenario": "high"}
-    {"source": ["FAOSTAT", "UNFCCC"]}
-    """
-
-    def match(self, fill_ts: xr.DataArray) -> bool:
-        """Check if a selected timeseries for filling matches this selector."""
-        return all(fill_ts.coords[k] == v for (k, v) in self.selections.items())
-
-    def match_single_dim(self, dim: Hashable, value: str) -> bool:
-        """Check if a literal value in one dimension can match this selector."""
-        if dim not in self.selections.keys():
-            return True
-        if isinstance(self.selections[dim], list):
-            return value in self.selections[dim]
-        return value == self.selections[dim]
-
-
-@define
-class StrategyDefinition:
-    """
-    Defines filling strategies for a single timeseries.
-
-    Attributes
-    ----------
-    strategies
-        List of mappings from a timeseries selector to a filling strategy. When a
-        timeseries will be used to fill missing data, the list will be checked from the
-        start, and the first matching TimeseriesSelector determines the FillingStrategy.
-        Example: [(("source", ["FAOSTAT", "UNFCCC]), StraightStrategy),
-                  ((, ), GlobalStrategy)]
-        Note that the strategy can depend on fixed coordinates as well as priority
-        coordinates.
-
-    """
-
-    strategies: list[tuple[TimeseriesSelector, FillingStrategyModel]]
-
-    def find_strategy(self, fill_ts: xr.DataArray) -> FillingStrategyModel:
-        """Find the strategy to use for the given filling timeseries."""
-        for selector, strategy in self.strategies:
-            if selector.match(fill_ts):
-                return strategy
-        raise KeyError(f"No matching strategy found for {fill_ts.coords}")
-
-    def limit(self, dim: Hashable, value: str) -> typing.Self:
-        """Limit this strategy definition to strategies applicable with the limit."""
-        return StrategyDefinition(
-            strategies=[
-                (sel, strat)
-                for (sel, strat) in self.strategies
-                if sel.match_single_dim(dim, value)
-            ]
+        result_das[f"Processing of {entity}"] = xr.DataArray(
+            data=np.empty(
+                [len(input_da.coords[dim]) for dim in group_by_dimensions], dtype=object
+            ),
+            dims=group_by_dimensions,
+            coords=[input_da.coords[dim] for dim in group_by_dimensions],
         )
+        number_of_timeseries = math.prod(
+            len(input_da[dim]) for dim in group_by_dimensions
+        )
+        if progress_bar:
+            pbar = tqdm.tqdm(total=number_of_timeseries, unit="ts", unit_scale=True)
+        else:
+            pbar = None
+        iterate_next_fixed_dimension(
+            input_da=input_da,
+            priority_definition=priority_definition,
+            strategy_definition=strategy_definition,
+            group_by_dimensions=group_by_dimensions,
+            result_da=result_das[entity],
+            result_processing_da=result_das[f"Processing of {entity}"],
+            progress_bar=pbar,
+        )
+        if progress_bar:
+            pbar.close()
+
+    return xr.Dataset(result_das).pr.quantify()
 
 
 def priority_coordinates_repr(
     *, fill_ts: xr.DataArray, priority_dimensions: list[str]
 ) -> str:
-    """Reduce the priority coordinates to a short string representation"""
+    """Reduce the priority coordinates to a short string representation."""
     priority_coordinates: dict[str, str] = {
         k: fill_ts[k].item() for k in priority_dimensions
     }
@@ -237,12 +142,59 @@ def priority_coordinates_repr(
     return repr(priority_coordinates)
 
 
+def iterate_next_fixed_dimension(
+    *,
+    input_da: xr.DataArray,
+    priority_definition: _models.PriorityDefinition,
+    strategy_definition: _models.StrategyDefinition,
+    group_by_dimensions: tuple[Hashable, ...],
+    result_da: xr.DataArray,
+    result_processing_da: xr.DataArray,
+    progress_bar: tqdm.tqdm | None,
+) -> None:
+    """Recursively iterate over dimensions in group_by_dimensions.
+
+    If there is only one dimension left, actually compute results and store
+    them in the result_da and the result_processing_da. Otherwise, iterate over one
+    dimension and recursively call iterate_next_fixed_dimension again.
+    """
+    my_dim = group_by_dimensions[0]
+    new_group_by_dimensions = group_by_dimensions[1:]
+    for val_array in input_da[my_dim]:
+        val = val_array.item()
+        strategy_definition = strategy_definition.limit(dim=my_dim, value=val)
+        priority_definition = priority_definition.limit(dim=my_dim, value=val)
+        if new_group_by_dimensions:
+            # have to iterate further until all dimensions are consumed
+            iterate_next_fixed_dimension(
+                input_da=input_da.loc[{my_dim: val}],
+                priority_definition=priority_definition,
+                strategy_definition=strategy_definition,
+                group_by_dimensions=new_group_by_dimensions,
+                result_da=result_da.loc[{my_dim: val}],
+                result_processing_da=result_processing_da.loc[{my_dim: val}],
+                progress_bar=progress_bar,
+            )
+        else:
+            # actually compute results
+            (
+                result_da.loc[{my_dim: val}],
+                result_processing_da.loc[{my_dim: val}],
+            ) = compose_timeseries(
+                input_data=input_da.loc[{my_dim: val}],
+                priority_definition=priority_definition,
+                strategy_definition=strategy_definition,
+            )
+            if progress_bar is not None:
+                progress_bar.update()
+
+
 def compose_timeseries(
     *,
     input_data: xr.DataArray,
-    priority_definition: PriorityDefinition,
-    strategy_definition: StrategyDefinition,
-) -> tuple[xr.DataArray, TimeseriesProcessingDescription]:
+    priority_definition: _models.PriorityDefinition,
+    strategy_definition: _models.StrategyDefinition,
+) -> tuple[xr.DataArray, _models.TimeseriesProcessingDescription]:
     """
     Compute a single timeseries from given input data, priorities, and strategies.
 
@@ -260,10 +212,10 @@ def compose_timeseries(
 
     Returns
     -------
-        result_ts, result_sources_ts. In result_ts is the numerical result, in
-        result_sources_ts is the string representation of the derivation of each data
-        point. Both are a single timeseries with the same fixed coordinates as
-        input_data and the same time dimension like input_data.
+        result_ts, processing_description. In result_ts is the numerical result, with
+        the time as the only dimension.
+        processing_description is the representation of the processing steps taken to
+        derive the result.
     """
     context_logger = logger.bind(
         fixed_coordinates={k: v for k, v in input_data.coords.items() if v.shape == ()},
@@ -299,7 +251,7 @@ def compose_timeseries(
                 f"basis to fill."
             )
             processing_steps_descriptions.append(
-                ProcessingStepDescription(
+                _models.ProcessingStepDescription(
                     time="all",
                     processing_description=f"used values from {fill_ts_repr}",
                 )
@@ -324,97 +276,6 @@ def compose_timeseries(
             f"No selector matched for \n{input_data.coords}\n{priority_definition=}"
         )
 
-    return result_ts, TimeseriesProcessingDescription(
+    return result_ts, _models.TimeseriesProcessingDescription(
         steps=processing_steps_descriptions
     )
-
-
-def compose(
-    *,
-    input_data: xr.Dataset,
-    priority_definition: PriorityDefinition,
-    strategy_definition: StrategyDefinition,
-) -> xr.Dataset:
-    result_das = {}
-
-    entity_pbar = tqdm.tqdm(input_data)
-    for entity in entity_pbar:
-        entity_pbar.set_postfix_str(entity)
-        input_da = input_data[entity].pr.dequantify()
-        # all dimensions are either time, priority selection dimensions, or need to
-        # be iterated over
-        group_by_dimensions = tuple(
-            dim
-            for dim in input_da.dims
-            if dim != "time" and dim not in priority_definition.selection_dimensions
-        )
-        # pre-allocate result arrays which will be filled timeseries-by-timeseries
-        result_dimensions = ["time", *group_by_dimensions]
-        result_das[entity] = xr.DataArray(
-            data=np.nan,
-            dims=result_dimensions,
-            coords=[input_da.coords[dim] for dim in result_dimensions],
-        )
-        result_das[f"Processing of {entity}"] = xr.DataArray(
-            data=np.empty(
-                [len(input_da.coords[dim]) for dim in group_by_dimensions], dtype=object
-            ),
-            dims=group_by_dimensions,
-            coords=[input_da.coords[dim] for dim in group_by_dimensions],
-        )
-        number_of_timeseries = math.prod(
-            len(input_da[dim]) for dim in group_by_dimensions
-        )
-        pbar = tqdm.tqdm(total=number_of_timeseries, unit="ts", unit_scale=True)
-        iterate_next_fixed_dimension(
-            input_da=input_da,
-            priority_definition=priority_definition,
-            strategy_definition=strategy_definition,
-            group_by_dimensions=group_by_dimensions,
-            result_da=result_das[entity],
-            result_processing_da=result_das[f"Processing of {entity}"],
-            progress_bar=pbar,
-        )
-        pbar.close()
-
-    return xr.Dataset(result_das).pr.quantify()
-
-
-def iterate_next_fixed_dimension(
-    *,
-    input_da: xr.DataArray,
-    priority_definition: PriorityDefinition,
-    strategy_definition: StrategyDefinition,
-    group_by_dimensions: tuple[Hashable, ...],
-    result_da: xr.DataArray,
-    result_processing_da: xr.DataArray,
-    progress_bar: tqdm.tqdm,
-) -> None:
-    my_dim = group_by_dimensions[0]
-    new_group_by_dimensions = group_by_dimensions[1:]
-    for val_array in input_da[my_dim]:
-        val = val_array.item()
-        strategy_definition = strategy_definition.limit(dim=my_dim, value=val)
-        priority_definition = priority_definition.limit(dim=my_dim, value=val)
-        if new_group_by_dimensions:
-            # have to iterate further until all dimensions are consumed
-            iterate_next_fixed_dimension(
-                input_da=input_da.loc[{my_dim: val}],
-                priority_definition=priority_definition,
-                strategy_definition=strategy_definition,
-                group_by_dimensions=new_group_by_dimensions,
-                result_da=result_da.loc[{my_dim: val}],
-                result_processing_da=result_processing_da.loc[{my_dim: val}],
-                progress_bar=progress_bar,
-            )
-        else:
-            # actually compute results
-            (
-                result_da.loc[{my_dim: val}],
-                result_processing_da.loc[{my_dim: val}],
-            ) = compose_timeseries(
-                input_data=input_da.loc[{my_dim: val}],
-                priority_definition=priority_definition,
-                strategy_definition=strategy_definition,
-            )
-            progress_bar.update()
